requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
nest-asyncio>=1.5.0
pyngrok>=7.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
grad-cam>=1.5.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
